Refactor additionalInfo field

+++ src/main/java/com/robertsoultanaev/javasphinx/packet/RelayInfo.java
@@ -1,4 +1,4 @@
 package com.robertsoultanaev.javasphinx.packet;
 
-public record RelayInfo(int nextNodeId, long additionalInfo) {
+public record RelayInfo(int nextNodeId, int additionalInfo) {
 }
